Auction with the winner's notarized
Final project on a study blockchain branch.
This is an asset sale auction. The primary features, compared to classic auctions, is the registration of the bids on the Redis db and the notarizaiton of the winner via the blockchain -Goerli testnet.
The auction can be closed either when the countdown ends or by closing it manually; when the auction closes don't forget to refresh and click on recorded operation. with createsuperuser you have the possibility to monitor all auctions with JSON.